+++ NEWS.md
@@ -9,6 +9,8 @@
 * Import `pillar::pillar_options` since it is mentioned in the documentation of `print.MarkovMix()`,
 as reported from platforms as notes that `pillar` is listed as "Imports" but no function is imported.
 
+* Fix incorrect documentation of `value` parameter in the replacement method of `MarkovMix` objects.
+
 
 # markovmix 0.1.1