+++ NEWS.md
@@ -4,12 +4,16 @@
 
 * Update package documentation.
 
+* Add documentation to C++ functions and export them to C++ interface.
+
 ## Bug fixes:
 
 * Import `pillar::pillar_options` since it is mentioned in the documentation of `print.MarkovMix()`,
 as reported from platforms as notes that `pillar` is listed as "Imports" but no function is imported.
 
 * Fix incorrect documentation of `value` parameter in the replacement method of `MarkovMix` objects.
+
+* Fix typos in the change logs of the documentations of `MarkovMix` utility functions.
 
 
 # markovmix 0.1.1